德州扑克

+++ src/main/java/com/example/dpoker/dto/GameUpdateDto.java
@@ -12,7 +12,7 @@ public class GameUpdateDto {
     private List<PlayerView> players;
     private List<String> communityCards; // e.g., ["Ah", "Kd", "2s"]
     private int currentBet;
-    private Long currentPlayerId;
+    private Integer currentPlayerId;
     private List<PotView> pots;
     private String currentRound;        // "PREFLOP", "FLOP", "TURN", "RIVER"
     private Map<String, Object> metadata; // 额外信息，如 winners、folded 玩家等
